requirements.txt
streamlit==1.31.0
oracledb==2.0.0
python-dotenv==1.0.0